fixed more imports; all tests pass
