+++ arkouda/numeric.py
@@ -5,9 +5,7 @@ from arkouda.dtypes import *
 from arkouda.pdarrayclass import pdarray, parse_single_value, create_pdarray
 from arkouda.pdarraysetops import unique
 
-__all__ = ["abs", "log", "exp", "cumsum", "cumprod", "sin", "cos",
-           "any", "all", "is_sorted", "sum", "prod", "min", "max",
-           "argmin", "argmax", "mean", "var", "std", "where",
+__all__ = ["abs", "log", "exp", "cumsum", "cumprod", "sin", "cos", "where",
            "histogram", "value_counts"]
 
 def abs(pda):
@@ -100,180 +98,6 @@ def cos(pda):
         return create_pdarray(repMsg)
     else:
         raise TypeError("must be pdarray {}".format(pda))
-
-def any(pda):
-    """
-    Return True iff any element of the array evaluates to True.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("any", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def all(pda):
-    """
-    Return True iff all elements of the array evaluate to True.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("all", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-    
-def is_sorted(pda):
-    """
-    Return True iff the array is monotonically non-decreasing.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("is_sorted", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def sum(pda):
-    """
-    Return the sum of all elements in the array.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("sum", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def prod(pda):
-    """
-    Return the product of all elements in the array. Return value is
-    always a float.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("prod", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def min(pda):
-    """
-    Return the minimum value of the array.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("min", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def max(pda):
-    """
-    Return the maximum value of the array.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("max", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-    
-def argmin(pda):
-    """
-    Return the index of the first minimum value of the array.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("argmin", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def argmax(pda):
-    """
-    Return the index of the first maximum value of the array.
-    """
-    if isinstance(pda, pdarray):
-        repMsg = generic_msg("reduction {} {}".format("argmax", pda.name))
-        return parse_single_value(repMsg)
-    else:
-        raise TypeError("must be pdarray {}".format(pda))
-
-def mean(pda):
-    """
-    Return the mean of the array.
-    """
-    return pda.sum() / pda.size
-
-def var(pda, ddof=0):
-    """
-    Return the variance of values in the array.
-
-    Parameters
-    ----------
-    pda : pdarray
-        Values for which to find the variance
-    ddof : int
-        "Delta Degrees of Freedom" used in calculating mean
-
-    Returns
-    -------
-    float
-        The scalar variance of the array
-
-    See Also
-    --------
-    mean, std
-
-    Notes
-    -----
-    The variance is the average of the squared deviations from the mean,
-    i.e.,  ``var = mean((x - x.mean())**2)``.
-
-    The mean is normally calculated as ``x.sum() / N``, where ``N = len(x)``.
-    If, however, `ddof` is specified, the divisor ``N - ddof`` is used
-    instead.  In standard statistical practice, ``ddof=1`` provides an
-    unbiased estimator of the variance of a hypothetical infinite population.
-    ``ddof=0`` provides a maximum likelihood estimate of the variance for
-    normally distributed variables.
-    """
-    if not isinstance(pda, pdarray):
-        raise TypeError("must be pdarray {}".format(pda))
-    if ddof >= pda.size:
-        raise ValueError("var: ddof must be less than number of values")
-    m = mean(pda)
-    return ((pda - m)**2).sum() / (pda.size - ddof)
-
-def std(pda, ddof=0):
-    """
-    Return the standard deviation of values in the array. The standard
-    deviation is implemented as the square root of the variance.
-
-    Parameters
-    ----------
-    pda : pdarray
-        values for which to find the variance
-    ddof : int
-        "Delta Degrees of Freedom" used in calculating mean
-
-    Returns
-    -------
-    float
-        The scalar standard deviation of the array
-
-    See Also
-    --------
-    mean, var
-
-    Notes
-    -----
-    The standard deviation is the square root of the average of the squared
-    deviations from the mean, i.e., ``std = sqrt(mean((x - x.mean())**2))``.
-
-    The average squared deviation is normally calculated as
-    ``x.sum() / N``, where ``N = len(x)``.  If, however, `ddof` is specified,
-    the divisor ``N - ddof`` is used instead. In standard statistical
-    practice, ``ddof=1`` provides an unbiased estimator of the variance
-    of the infinite population. ``ddof=0`` provides a maximum likelihood
-    estimate of the variance for normally distributed variables. The
-    standard deviation computed in this function is the square root of
-    the estimated variance, so even with ``ddof=1``, it will not be an
-    unbiased estimate of the standard deviation per se.
-    """
-    return np.sqrt(var(pda, ddof=ddof))
     
 def where(condition, A, B):
     """
